rooms: port Room_AlterFloorHeight to TRX
