+++ src/libtrx/include/libtrx/game/rooms/geometry.h
@@ -23,7 +23,7 @@ int16_t Room_GetCeilingEx(
     const SECTOR *sector, int32_t x, int32_t y, int32_t z, bool fix_tilts);
 
 int32_t Room_GetWaterHeight(int32_t x, int32_t y, int32_t z, int16_t room_num);
-extern void Room_AlterFloorHeight(const ITEM *item, int32_t height);
+void Room_AlterFloorHeight(const ITEM *item, int32_t height);
 
 extern int32_t Room_FindGridShift(int32_t src, int32_t dst);
